+++ widget/accordion.go
@@ -137,8 +137,8 @@ type accordionRenderer struct {
 }
 
 func (r *accordionRenderer) Layout(size fyne.Size) {
-	x := 0
-	y := 0
+	x := float32(0)
+	y := float32(0)
 	for i, ai := range r.container.Items {
 		if i != 0 {
 			div := r.dividers[i-1]
